No Bug - Consolidated testing Deferred extensions into it's own, shared source
